+++ kinetic_datanator/api/urls.py
@@ -1,8 +1,6 @@
 from kinetic_datanator.api.views import *
 
 V0_ENDPOINT = '/v0'
-
-
 # Text Search
 api.add_resource(Search, V0_ENDPOINT+'/search/<value>')
