+++ kinetic_datanator/core/models.py
@@ -4,7 +4,7 @@ from sqlalchemy_searchable import SearchQueryMixin, make_searchable
 from sqlalchemy_utils.types import TSVectorType
 from flask_migrate import Migrate
 from kinetic_datanator.config import config
-from kinetic_datanator.app import create_app, db
+from kinetic_datanator import db
 from flask_sqlalchemy import SQLAlchemy
 import os
